Moved the databases config stuff into its one directory
